requirements.txt
Babel==2.6.0
Django==1.11.15
django-language-field==0.0.3
django-money==0.14.3
django-phonenumber-field==2.0.1
django-tastypie==0.14.1
djangorestframework==3.8.2
djangorestframework-gis==0.13
phonenumbers==8.9.11
pkg-resources==0.0.0
psycopg2==2.7.5
py-moneyed==0.7.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.5
six==1.11.0
